refactor(service): change the logic of saving to a file
The logic of saving to a file has been changed. In order
not to clog up the business layer, the logic of saving
to a file has been moved to a separate decorator. Saving
to a file is a specific event only for using the inmemory repository.
The problem was that because of this dependency, all the tests
had to be changed. When using a real database, the fs and
StoreInterval parameters clogged up the service signature and
created unnecessary dependencies on data storage, that is, the
infrastructure layer.